lean version & improve db version
