--- test-clients/amq-consumer/src/main/resources/application.properties
@@ -1,21 +1,0 @@
-quarkus.http.port=9191
-
-quarkus.log.level=INFO
-
-#The AMQP broker location and credentials
-amqp-host=163.66.93.15
-amqp-port=5672
-amqp-username=admin
-amqp-password=redhat
-
-#Configuring the incoming channel (reading from AMQP)
-mp.messaging.incoming.exampleQueue-in.connector=smallrye-amqp
-mp.messaging.incoming.exampleQueue-in.address=SampleQueue
-
-#Configuring the outgoing channel (writing to AMQP)
-mp.messaging.outgoing.exampleQueue-out.connector=smallrye-amqp
-mp.messaging.outgoing.exampleQueue-out.address=SampleQueue
-mp.messaging.outgoing.exampleQueue-out.durable=false
-
-quarkus.datasource.db-kind=h2
-quarkus.datasource.jdbc.url=jdbc:h2:file:~/temp/h2/amq/main;AUTO_SERVER=true;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=FALSE
