+++ test-clients/db-version/amq-producer/src/main/java/com/redhat/amq/benchmark/entity/TestMetrics.java
@@ -1,0 +1,22 @@
+package com.redhat.amq.benchmark.entity;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
+
+import lombok.Data;
+
+import java.sql.Timestamp;
+import java.time.LocalDateTime;
+
+@JsonAutoDetect(fieldVisibility = Visibility.ANY)
+@JsonPropertyOrder
+@Data
+public class TestMetrics {
+
+    private long elapsedTimeMillis;
+    private long totalMessagesSent;
+    private String message;
+    private Timestamp producerStartTime;
+    private Timestamp producerEndTime;
+}
