+++ test-clients/db-version/amq-producer/src/main/java/com/redhat/amq/benchmark/services/MessageProducerService.java
@@ -1,0 +1,33 @@
+package com.redhat.amq.benchmark.services;
+
+import com.redhat.amq.benchmark.resources.MessageResource;
+import org.eclipse.microprofile.reactive.messaging.Channel;
+import org.eclipse.microprofile.reactive.messaging.Emitter;
+import org.eclipse.microprofile.reactive.messaging.OnOverflow;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+
+import com.redhat.amq.benchmark.entity.Message;
+
+import java.sql.Timestamp;
+import java.time.Instant;
+
+@ApplicationScoped
+public class MessageProducerService {
+
+    @Inject
+    @Channel("exampleQueue-out")
+    @OnOverflow(value = OnOverflow.Strategy.UNBOUNDED_BUFFER)
+    Emitter<Message> emitter;
+
+    @Inject
+    MessageResource messageResource;
+
+    public Message send(Message message) {
+        message.setSent(Timestamp.from(Instant.now()));
+        emitter.send(message);
+        messageResource.create(message);
+        return message;
+    }
+}
