+++ test-clients/no-db-lean-version/amq-consumer/src/main/java/com/redhat/amq/benchmark/pojo/Message.java
