+++ test-clients/no-db-lean-version/amq-consumer/src/main/resources/application.properties
@@ -1,0 +1,13 @@
+quarkus.http.port=9191
+
+quarkus.log.level=INFO
+
+#The AMQP broker location and credentials
+amqp-host=163.66.93.15
+amqp-port=5672
+amqp-username=admin
+amqp-password=redhat
+
+#Configuring the incoming channel (reading from AMQP)
+mp.messaging.incoming.exampleQueue-in.connector=smallrye-amqp
+mp.messaging.incoming.exampleQueue-in.address=SampleQueue
