+++ test-clients/no-db-lean-version/amq-producer/src/main/java/com/redhat/amq/benchmark/pojo/Message.java
