+++ test-clients/no-db-lean-version/amq-producer/src/main/resources/application.properties
@@ -1,0 +1,14 @@
+quarkus.http.port=9090
+
+quarkus.log.level=INFO
+
+#The AMQP broker location and credentials
+amqp-host=163.66.93.15
+amqp-port=5672
+amqp-username=amq
+amqp-password=redhat
+
+#Configuring the outgoing channel (writing to AMQP)
+mp.messaging.outgoing.exampleQueue-out.connector=smallrye-amqp
+mp.messaging.outgoing.exampleQueue-out.address=SampleQueue
+mp.messaging.outgoing.exampleQueue-out.durable=false
